int to float

+++ pkg/math_client/math_client_interfaces.go
@@ -1,5 +1,5 @@
 package pkg
 
 type MathGrpcClient interface {
-	Calculate(x int64, y int64, operationSign string, result *int64) error
+	Calculate(x float64, y float64, operationSign string, result *float64) error
 }
